requirements.txt
fastapi
uvicorn
python-multipart
Pillow
PyMuPDF
jinja2
